requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
eval-type-backport==0.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
invoke==2.2.1
mistralai==1.9.11
pdf2image==1.17.0
pillow==12.0.0
pydantic==2.12.5
pydantic-core==2.41.5
pypdf==6.4.0
python-dateutil==2.9.0.post0
pyyaml==6.0.3
six==1.17.0
sniffio==1.3.1
typing-extensions==4.15.0
typing-inspection==0.4.2
